Atualização para funcionar com o novo banco de dados

+++ src/mercadinho/Classes/Forn.php
@@ -6,7 +6,7 @@ class Forn
     }
 
     function insereFornecedor($nome, $telefone, $email, $estado, $cidade, $endereco, $conn){
-        $sql = "INSERT INTO fornecedores (nome, telefone, email, estado, cidade, endereço) 
+        $sql = "INSERT INTO fornecedores (nome, telefone, email, estado, cidade, endereço)
         VALUES ('$nome','$telefone', '$email', '$estado', '$cidade', '$endereco')";
 
         if ($conn->query($sql) === TRUE) {
@@ -18,12 +18,12 @@ class Forn
     }
 
     function removeFornecedor($nome, $conn){
-        $sql = "SELECT id, nome, telefone, email, estado, cidade, endereço 
+        $sql = "SELECT id_fornecedor, nome, telefone, email, estado, cidade, endereço
             FROM fornecedores
             WHERE nome='$nome'";
         $result = $conn->query($sql);
-        if ($result->num_rows > 0){
-            $sql = "DELETE FROM fornecedores 
+        if ($result and $result->num_rows > 0) {
+            $sql = "DELETE FROM fornecedores
         WHERE nome = '$nome'";
 
             if ($conn->query($sql) === TRUE) {
@@ -53,15 +53,15 @@ class Forn
 
     function listarFornecedor($nome, $conn){
         if($nome == null){
-            $sql = "SELECT id, nome, telefone, email, estado, cidade, endereço 
+            $sql = "SELECT id_fornecedor, nome, telefone, email, estado, cidade, endereço
             FROM fornecedores";
             $result = $conn->query($sql);
 
             echo "<B>Lista de Fornecedores:</B> <br>";
-            if ($result->num_rows > 0) {
+            if ($result and $result->num_rows > 0) {
                 // output data of each row
                 while($row = $result->fetch_assoc()) {
-                    echo "<b>id:</b> " . $row["id"]. " - <b>Nome:</b> " . $row["nome"]. " - <b>Telefone:</b> " . $row["telefone"].
+                    echo "<b>id:</b> " . $row["id_fornecedor"]. " - <b>Nome:</b> " . $row["nome"]. " - <b>Telefone:</b> " . $row["telefone"].
                         " -<b> Email:</b> " . $row["email"]. " -<b> Estado:</b> " . $row["estado"]. " - <b>Cidade:</b> " . $row["cidade"].
                         " - <b>endereço:</b> " . $row["endereço"]."<br>";
                 }
@@ -73,16 +73,15 @@ class Forn
 
         else{
             echo "<b>Resultado da busca:</b> <br>";
-            $sql = "SELECT id, nome, telefone, email, estado, cidade, endereço 
+            $sql = "SELECT id_funcionario, nome, telefone, email, estado, cidade, endereço
             FROM fornecedores
             WHERE nome='$nome'";
             $result = $conn->query($sql);
 
-            echo "<B>Resultado da busca:</B> <br>";
-            if ($result->num_rows > 0) {
+            if ($result and $result->num_rows > 0) {
                 // output data of each row
                 while($row = $result->fetch_assoc()) {
-                    echo "<b>id:</b> " . $row["id"]. " - <b>Nome:</b> " . $row["nome"]. " - <b>Telefone:</b> " . $row["telefone"].
+                    echo "<b>id:</b> " . $row["id_funcionario"]. " - <b>Nome:</b> " . $row["nome"]. " - <b>Telefone:</b> " . $row["telefone"].
                         " -<b> Email:</b> " . $row["email"]. " -<b> Estado:</b> " . $row["estado"]. " - <b>Cidade:</b> " . $row["cidade"].
                         " - <b>endereço:</b> " . $row["endereço"]."<br>";
                 }
