+++ src/mercadinho/Classes/Prod.php
@@ -6,7 +6,7 @@ class Prod
     }
 
     function adicionarProduto($nome, $valor, $conn){
-        $sql = "INSERT INTO produtos (nome, valor) 
+        $sql = "INSERT INTO produtos (nome, valor)
         VALUES ('$nome','$valor')";
 
         if ($conn->query($sql) === TRUE) {
@@ -19,12 +19,12 @@ class Prod
 
     function removeProduto($nome, $conn)
     {
-        $sql = "SELECT id, nome, valor 
+        $sql = "SELECT id_produto, nome, valor
             FROM produtos
             WHERE nome = '$nome'";
         $result = $conn->query($sql);
-        if ($result->num_rows > 0) {
-            $sql = "DELETE FROM produtos 
+        if ($result and $result->num_rows > 0) {
+            $sql = "DELETE FROM produtos
         WHERE nome = '$nome'";
 
             if ($conn->query($sql) === TRUE) {
@@ -38,7 +38,7 @@ class Prod
     function editarProduto($nome, $novo_nome, $novo_valor, $conn){
 
         $sql = "UPDATE produtos
-        SET nome = '$novo_nome', valor = '$novo_valor' 
+        SET nome = '$novo_nome', valor = '$novo_valor'
         WHERE nome = '$nome'";
 
         if ($conn->query($sql) === TRUE) {
@@ -51,15 +51,15 @@ class Prod
 
     function listarProduto($nome, $conn){
         if($nome == null){
-            $sql = "SELECT id, nome, valor 
+            $sql = "SELECT id_produto, nome, valor
             FROM produtos";
             $result = $conn->query($sql);
 
-            if ($result->num_rows > 0) {
+            if ($result and $result->num_rows > 0) {
                 // output data of each row
                 echo "<B>Lista de produtos</B> <br>";
                 while($row = $result->fetch_assoc()) {
-                    echo "<b>id:</b> " . $row["id"]. " - <b>Nome:</b> " . $row["nome"]. " - <b>Valor:</b> R$ " . $row["valor"]. "<br>";
+                    echo "<b>id:</b> " . $row["id_produto"]. " - <b>Nome:</b> " . $row["nome"]. " - <b>Valor:</b> R$ " . $row["valor"]. "<br>";
                 }
             }
             else {
@@ -69,15 +69,15 @@ class Prod
 
         else{
             echo "<b>Resultado da busca:</b> <br>";
-            $sql = "SELECT id, nome, valor 
+            $sql = "SELECT id_produto, nome, valor
             FROM produtos
             WHERE nome = '$nome'";
             $result = $conn->query($sql);
 
-            if ($result->num_rows > 0) {
+            if ($result and $result->num_rows > 0) {
                 // output data of each row
                 while($row = $result->fetch_assoc()) {
-                    echo "<b>id:</b> " . $row["id"]. " - <b>Nome:</b> " . $row["nome"]. " - <b>Valor:</b> R$ " . $row["valor"]. "<br>";
+                    echo "<b>id:</b> " . $row["id_produto"]. " - <b>Nome:</b> " . $row["nome"]. " - <b>Valor:</b> R$ " . $row["valor"]. "<br>";
                 }
             }
             else {
